examples/marketing-campaign/pyproject.toml
[project]
name = "marketing-campaign"
version = "0.1.0"
description = ""
authors = [
    {name = "Alessandro Duminuco",email = "aduminuc@cisco.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dynamic = [ "dependencies" ]

[tool.poetry.dependencies]
agntcy-iomapper = { path = "../../../iomapper-agnt", develop = true}
acp-sdk = { path = "../../acp-sdk", develop = true}
python-dotenv = "^1.0.1"
langgraph = "^0.3.5"
langchain-openai = "^0.3.8"


[tool.poetry]
packages = [{include = "marketing_campaign", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
